GitVerse Tasks/GitVerse_Task_1/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sysaudit"
version = "0.1.0"
description = "Git-based System Audit & Compliance Monitor"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "System Audit Team"}
]
keywords = ["audit", "compliance", "security", "monitoring", "git"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "Topic :: Security",
    "Topic :: System :: Monitoring",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: POSIX :: Linux",
]

dependencies = [
    "watchdog>=3.0.0",
    "GitPython>=3.1.0",
    "click>=8.0.0",
    "PyYAML>=6.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
sysaudit = "sysaudit.cli:main"

[project.urls]
Homepage = "https://github.com/sysaudit/sysaudit"
Documentation = "https://github.com/sysaudit/sysaudit/docs"
Repository = "https://github.com/sysaudit/sysaudit"
Issues = "https://github.com/sysaudit/sysaudit/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["sysaudit*"]
exclude = ["tests*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=sysaudit --cov-report=term-missing"
